requirements.txt
Flask-login
Werkzeug
itsdangerous
Flask-Bootstrap
Flask-WTF
Flask-script
Flask-Mail
Flask-migrate
pymysql